Decode Unicode in XML requests earlier.
Previously, Unicode in input requests was only decoded when parsing the
XML. However, requests are already processed before that.

Unicode is now decoded at an earlier point, so that requests are saved
to the database with proper encoding. This means that the behaviour of
`parseXML` changes: passing improperly encoded Unicode to it will now
result in gibberish.

Relates to the ticket http://ideastest.science.uu.nl/trac/ticket/240